Introduce actor for wifi_user
