+++ spec/unit/lib/common/base64_spec.rb
