+++ spec/unit/lib/performance_platform/gateway/performance_report_spec.rb
