+++ spec/unit/lib/performance_platform/use_case/send_performance_report_spec.rb
